Replaced observables data types; Moved all hooks constants in separate file ✨✨✨
